fix(kreivo-runtime): adjust settings for `memberships` benchmarking
